Update the UI
This adds the carousel for showing the player the instructions. This
updates the styles for the welcome screen and high score screen. Some
of the sytles on the carousel still need to be set and scrolling and
layout need to updated for smaller screens.

The system needed to be updated to handle the case where there are no
currently set high scores.